+++ src/utils/img_detect_ai.py
@@ -48,12 +48,14 @@ def detect_img_py_v1(img_path):
                     result = "other"
                 return result + ": " + str(max_score)
             else:
-                logger.warning(f"unknown error, detail: {img_path}, please re click AI detect img")
-                constants.detect_model_flag = False
+                logger.warning(f"unknown error, server response: {response}, detail: {img_path}, please re click AI "
+                               f"detect img")
+                # constants.detect_model_flag = False
                 return False
     except Exception as e:
-        logger.warning(f"unknown error, detail: {e}, please re click AI detect img")
-        constants.detect_model_flag = False
+        logger.warning(f"unknown error, server response: {response}, detail: {e}, error image_path: {img_path}, please "
+                       f"re click AI detect img")
+        # constants.detect_model_flag = False
         return False
 
 
@@ -157,12 +159,14 @@ def model_detect_img_java_v1(img_path):
                     result = "other"
                 return result + ": " + str(max_score)
             else:
-                logger.warning(f"unknown error, detail: {img_path}, please re click AI detect img")
-                constants.detect_model_flag = False
+                logger.warning(f"unknown error, server response: {response}, detail: {img_path}, please re click AI "
+                               f"detect img")
+                # constants.detect_model_flag = False
                 return False
     except Exception as e:
-        logger.warning(f"unknown error, detail: {e}, please re click AI detect img")
-        constants.detect_model_flag = False
+        logger.warning(f"unknown error, server response: {response}, detail: {e}, error image_path: {img_path}, please "
+                       f"re click AI detect img")
+        # constants.detect_model_flag = False
         return False